requirements.txt
# Web framework
Flask>=2.2.2
Flask-PyMongo>=2.3.0
Flask-WTF>=1.1.1
werkzeug>=2.2.0

# Authentication
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
PyJWT>=2.7.0
bcrypt>=4.0.1

# Caching (instead of pip._vendor.cachecontrol)
cachecontrol>=0.13.0

# Document processing
PyMuPDF>=1.21.1
python-docx>=0.8.11
docx2txt>=0.8
python-magic>=0.4.27

# Image processing
Pillow>=9.0.0

# Database
pymongo>=4.0.0
redis>=4.5.0

# ML/Data Science - Compatible versions
torch>=1.13.0
transformers>=4.21.0
spacy>=3.5.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# NLP utilities
nltk>=3.8.0

# Utilities
joblib>=1.3.0
networkx>=3.1
tokenizers>=0.13.0
apscheduler>=3.9.0
python-dotenv>=1.0.0


# Email validation
email-validator>=2.0.0
